Remove dependencies from LoggingFormatStringHelpers
