+++ src/Access/AccessChangesNotifier.h
@@ -3,6 +3,7 @@
 #include <Access/IAccessEntity.h>
 #include <base/scope_guard.h>
 #include <list>
+#include <mutex>
 #include <queue>
 #include <unordered_map>
