+++ src/Access/Common/AccessType.h
@@ -5,6 +5,8 @@
 #include <boost/algorithm/string/split.hpp>
 #include <boost/algorithm/string/trim.hpp>
 
+#include <map>
+
 namespace DB
 {
