+++ src/Backups/BackupsWorker.h
@@ -6,6 +6,8 @@
 #include <Interpreters/Context_fwd.h>
 #include <Core/UUID.h>
 #include <Parsers/IAST_fwd.h>
+
+#include <mutex>
 #include <unordered_map>
