+++ src/Common/NamedCollections/NamedCollections.h
@@ -4,6 +4,8 @@
 #include <Parsers/ASTCreateNamedCollectionQuery.h>
 #include <Parsers/ASTAlterNamedCollectionQuery.h>
 
+#include <mutex>
+
 
 namespace Poco { namespace Util { class AbstractConfiguration; } }
