+++ src/Functions/UUIDv7Utils.h
@@ -2,6 +2,8 @@
 #include <Common/SharedMutex.h>
 #include <DataTypes/DataTypeUUID.h>
 
+#include <mutex>
+
 /// Common functionality for UUIDv7-related functions.
 
 namespace DB
