+++ src/Interpreters/Aggregator.h
@@ -10,6 +10,7 @@
 #include <Core/Block.h>
 #include <Core/Block_fwd.h>
 #include <Core/ColumnNumbers.h>
+#include <Common/Logger.h>
 #include <Common/ThreadPool.h>
 #include <Common/filesystemHelpers.h>
