+++ src/Interpreters/Cache/QueryConditionCache.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <Common/CacheBase.h>
+#include <Common/Logger.h>
 #include <Storages/MergeTree/MarkRange.h>
 #include <Common/SharedMutex.h>
