+++ src/Planner/Planner.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <Common/Logger.h>
+
 #include <Interpreters/IInterpreter.h>
 #include <Interpreters/SelectQueryOptions.h>
