+++ src/Processors/Formats/InputFormatErrorsLogger.h
@@ -5,6 +5,8 @@
 
 #include <Core/Block.h>
 
+#include <mutex>
+
 
 namespace DB
 {
