+++ src/Processors/Transforms/CreatingSetsTransform.h
@@ -6,6 +6,7 @@
 #include <QueryPipeline/Chain.h>
 #include <QueryPipeline/QueryPipeline.h>
 #include <Interpreters/PreparedSets.h>
+#include <Common/Logger.h>
 #include <Common/Stopwatch.h>
 
 #include <Poco/Logger.h>
