+++ src/Server/ProxyV1Handler.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <Common/Logger.h>
 #include <Poco/Net/TCPServerConnection.h>
 #include <Server/IServer.h>
 #include <Server/TCPProtocolStackData.h>
