+++ src/Storages/Distributed/DistributedAsyncInsertSource.cpp
@@ -1,10 +1,10 @@
 #include <Storages/Distributed/DistributedAsyncInsertSource.h>
 #include <Storages/Distributed/DistributedAsyncInsertHeader.h>
 #include <IO/ReadBufferFromFile.h>
+#include <Common/Logger.h>
 #include <Compression/CompressedReadBuffer.h>
 #include <Core/Settings.h>
 #include <Formats/NativeReader.h>
-#include <Poco/Logger.h>
 
 namespace DB
 {
