+++ src/Storages/Kafka/KafkaProducer.cpp
@@ -2,7 +2,7 @@
 #include <Core/Block.h>
 #include <Columns/ColumnString.h>
 #include <Columns/ColumnsNumber.h>
-
+#include <Common/Logger.h>
 #include <Common/ProfileEvents.h>
 
 namespace ProfileEvents
