+++ src/Storages/ObjectStorage/DataLakes/Iceberg/ComplexTypeSchemaProcessorFunctions.h
@@ -16,6 +16,8 @@
 #include <Poco/JSON/Stringifier.h>
 #include <Poco/SharedPtr.h>
 
+#include <mutex>
+
 namespace DB::Iceberg
 {
